requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
Authlib==1.6.6
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.4
cycler==0.12.1
distro==1.9.0
edge-tts==7.2.7
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
fonttools==4.58.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
MarkupSafe==3.0.3
matplotlib==3.10.3
multidict==6.7.0
narwhals==2.13.0
nest-asyncio==1.6.0
numpy==2.2.5
openai==2.9.0
packaging==25.0
pandas==2.3.3
pillow==11.2.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.5.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.52.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
watchdog==6.0.0
Werkzeug==3.1.5
yarl==1.22.0
